# answers_storage-openapi.yaml
openapi: "3.1.0"
info:
  title: "answers_storage API"
  description: "answers_storage API"
  version: "1.0.0"
servers:
  - url: "https://answers_storage"
paths:
  /get/all/text_match:
    get:
      summary: "GET get/all/text_match"
      operationId: "findByExerciseMatch"
      parameters:
        - name: "match"
          in: "query"
          required: true
          schema:
            type: "string"
        - name: "type"
          in: "query"
          required: true
          schema:
            type: "string"
            enum:
              - "PHYSICS"
              - "BIOLOGY"
              - "RUSSIAN"
              - "LITERATURE"
              - "SOCIOLOGY"
              - "INFORMATICS"
              - "HISTORY"
              - "OBJ"
              - "ENGLISH"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "string"
...
  /get/{id}:
    get:
      summary: "GET get/{id}"
      operationId: "getById"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "string"
...